Upgrade Android API target version to 28
Google Play Store will only accept apps targeting at least version 28 in
November 2019.

Set the minimum API version to 22. This version is for Android 5.1,
which a few of our users are still on.

The HttpClient API is deprecated but can still be used through a
uses-library instruction in the manifest.